Move AbstractFeatureTest to ops2 base dir
The net.imagej.ops2.features package exists only in the test scope, thus
it is not exported in the source module-info.java. This made tests
using that package (any feature test) unusable in Eclipse, for example.

This commit works around the issue by moving the AbstractFeatureTest to
the base net.imagej.ops2 directory, which is already being exported.
