+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/geom/geom2d/PolygonFeatureTests.java
@@ -32,7 +32,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import java.util.List;
 
-import net.imagej.ops2.features.AbstractFeatureTest;
+import net.imagej.ops2.AbstractFeatureTest;
 import net.imagej.ops2.geom.GeomUtils;
 import net.imglib2.RealLocalizable;
 import net.imglib2.RealPoint;
